Ok, I came this far.
I need to come up with a plan of learning. I cannot just copy paste the stuff from the SM100 example.

How about this:

* Create a Cute tensor on host and initialize it with some data.
* Pass it into a kernel and print it out on the device.

This is the most basic of the basics. I'll learn a lot by doing this. Let me read the Cute docs. I know they have Layout, Shape, Stride and Tensor types. But the example is also using Thrust, so I am not sure what I should be doing.

It needn't be a large tensor. A 5 element tensor is enough.